refactor(test): merge wandb config tests into test_configs.py
Move 3 wandb env var resolution tests from standalone
test_wandb_integration.py into test_configs.py where config tests
belong. Fix docstring to link issue #265 instead of opaque "Task 1.2".

Refs #265, refs #375
