+++ tests/test_configs.py
@@ -1,6 +1,6 @@
 import hydra
 from hydra.core.hydra_config import HydraConfig
-from omegaconf import DictConfig
+from omegaconf import DictConfig, OmegaConf
 
 
 def test_train_config(cfg_train: DictConfig) -> None:
@@ -35,3 +35,27 @@ def test_eval_config(cfg_eval: DictConfig) -> None:
     hydra.utils.instantiate(cfg_eval.data)
     hydra.utils.instantiate(cfg_eval.model)
     hydra.utils.instantiate(cfg_eval.trainer)
+
+
+class TestWandbConfigResolvesFromEnv:
+    """Verify wandb entity/project resolve from env vars (#265)."""
+
+    def test_wandb_entity_resolves_from_env(self, monkeypatch):
+        """OmegaConf resolves WANDB_ENTITY from environment."""
+        monkeypatch.setenv("WANDB_ENTITY", "test-entity")
+        cfg = OmegaConf.load("configs/logger/wandb.yaml")
+        assert OmegaConf.select(cfg, "wandb.entity") == "test-entity"
+
+    def test_wandb_project_resolves_from_env(self, monkeypatch):
+        """OmegaConf resolves WANDB_PROJECT from environment."""
+        monkeypatch.setenv("WANDB_PROJECT", "test-project")
+        cfg = OmegaConf.load("configs/logger/wandb.yaml")
+        assert OmegaConf.select(cfg, "wandb.project") == "test-project"
+
+    def test_wandb_defaults_when_env_unset(self, monkeypatch):
+        """Falls back to tinaudio/synth-setter when env vars unset."""
+        monkeypatch.delenv("WANDB_ENTITY", raising=False)
+        monkeypatch.delenv("WANDB_PROJECT", raising=False)
+        cfg = OmegaConf.load("configs/logger/wandb.yaml")
+        assert OmegaConf.select(cfg, "wandb.entity") == "tinaudio"
+        assert OmegaConf.select(cfg, "wandb.project") == "synth-setter"
